# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce REST API",
    "version": "1.0.0",
    "description": "A fully featured E-Commerce backend REST API built with Node.js, Express and MongoDB. Features include JWT Authentication, Product Management, Cart, Orders and Likes.",
    "contact": {
      "name": "Mrutunjaya Panda",
      "url": "https://github.com/Mrutunjaya-Panda"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3200",
      "description": "Local Development Server"
    },
    {
      "url": "https://e-comm-project-using-rest-api.onrender.com",
      "description": "Production Server (Render)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey",
        "description": "Enter your JWT token directly (without 'Bearer' prefix). Obtain a token from POST /api/users/signin."
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9a123"
          },
          "name": {
            "type": "string",
            "example": "Wireless Headphones"
          },
          "description": {
            "type": "string",
            "example": "Premium noise-cancelling over-ear wireless headphones."
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 2999.99
          },
          "inStock": {
            "type": "integer",
            "example": 50
          },
          "imageUrl": {
            "type": "string",
            "example": "1691234567890-headphones.jpg"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9b456"
          },
          "name": {
            "type": "string",
            "example": "Electronics"
          },
          "products": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["64b1f4c2e4b0f5a3d8e9a123"]
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9c789"
          },
          "productId": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9a123"
          },
          "userId": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9d000"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 4
          }
        }
      },
...
      "CartItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9e111"
          },
          "productId": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9a123"
          },
          "userId": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9d000"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "example": 2
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9f222"
          },
          "userId": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9d000"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          },
          "totalAmount": {
            "type": "number",
            "format": "float",
            "example": 5999.98
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "shipped", "delivered", "cancelled"],
            "example": "pending"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-08-30T06:17:00.000Z"
          }
        }
      },
      "Like": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9g333"
          },
          "user": {
            "type": "string",
            "example": "64b1f4c2e4b0f5a3d8e9d000"
          },
          "likeable": {
            "type": "string",
            "description": "ObjectId of the liked Product or Category",
            "example": "64b1f4c2e4b0f5a3d8e9a123"
          },
          "types": {
            "type": "string",
            "enum": ["Product", "Category"],
            "example": "Product"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "An error occurred."
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "description": "Retrieves a list of all products available in the store.",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "A list of all products returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                },
                "example": [
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9a123",
                    "name": "Wireless Headphones",
                    "description": "Premium noise-cancelling over-ear wireless headphones.",
                    "price": 2999.99,
                    "inStock": 50,
                    "imageUrl": "1691234567890-headphones.jpg",
                    "categories": ["64b1f4c2e4b0f5a3d8e9b456"],
                    "reviews": []
                  },
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9a124",
                    "name": "Running Shoes",
                    "description": "Lightweight and breathable running shoes for all terrains.",
                    "price": 1499.50,
                    "inStock": 120,
                    "imageUrl": "1691234567891-shoes.jpg",
                    "categories": ["64b1f4c2e4b0f5a3d8e9b457"],
                    "reviews": []
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Add a new product",
        "description": "Creates a new product with an optional image upload using multipart/form-data.",
        "security": [{ "JWT": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the product",
                    "example": "Wireless Headphones"
                  },
                  "description": {
                    "type": "string",
                    "description": "Detailed description of the product",
                    "example": "Premium noise-cancelling over-ear wireless headphones."
                  },
                  "price": {
                    "type": "number",
                    "description": "Price of the product",
                    "example": 2999.99
                  },
                  "categories": {
                    "type": "string",
                    "description": "Comma-separated list of category names",
                    "example": "Electronics, Audio"
                  },
                  "sizes": {
                    "type": "string",
                    "description": "Comma-separated list of available sizes (if applicable)",
                    "example": "S, M, L, XL"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image file to upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" },
                "example": {
                  "_id": "64b1f4c2e4b0f5a3d8e9a123",
                  "name": "Wireless Headphones",
                  "description": "Premium noise-cancelling over-ear wireless headphones.",
                  "price": 2999.99,
                  "imageUrl": "1691234567890-headphones.jpg",
                  "categories": [],
                  "reviews": []
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation failed or missing required fields.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Product validation failed: name: Path `name` is required." }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      }
    },
...
    "/api/cartItems": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get cart items for the authenticated user",
        "description": "Returns all cart items belonging to the currently authenticated user.",
        "security": [{ "JWT": [] }],
        "responses": {
          "200": {
            "description": "Cart items returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CartItem" }
                },
                "example": [
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9e111",
                    "productId": "64b1f4c2e4b0f5a3d8e9a123",
                    "userId": "64b1f4c2e4b0f5a3d8e9d000",
                    "quantity": 2
                  },
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9e112",
                    "productId": "64b1f4c2e4b0f5a3d8e9a124",
                    "userId": "64b1f4c2e4b0f5a3d8e9d000",
                    "quantity": 1
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Place an order",
        "description": "Places a new order for the authenticated user, converting their cart items into a confirmed order.",
        "security": [{ "JWT": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartItemIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Optional list of specific cart item ObjectIds to order. If omitted, all cart items are ordered.",
                    "example": [
                      "64b1f4c2e4b0f5a3d8e9e111",
                      "64b1f4c2e4b0f5a3d8e9e112"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" },
                "example": {
                  "_id": "64b1f4c2e4b0f5a3d8e9f222",
                  "userId": "64b1f4c2e4b0f5a3d8e9d000",
                  "items": [
                    {
                      "_id": "64b1f4c2e4b0f5a3d8e9e111",
                      "productId": "64b1f4c2e4b0f5a3d8e9a123",
                      "userId": "64b1f4c2e4b0f5a3d8e9d000",
                      "quantity": 2
                    }
                  ],
                  "totalAmount": 5999.98,
                  "status": "pending",
                  "createdAt": "2024-08-30T06:17:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - cart is empty or invalid cart item IDs provided.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "No items in cart to place an order." }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      }
    },
    "/api/likes": {
      "post": {
        "tags": ["Likes"],
        "summary": "Like a product or category",
        "description": "Allows an authenticated user to like a product or category by providing its ObjectId and entity type.",
        "security": [{ "JWT": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "type"],
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "The MongoDB ObjectId of the product or category to like.",
                    "example": "64b1f4c2e4b0f5a3d8e9a123"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["Product", "Category"],
                    "description": "The type of the entity being liked.",
                    "example": "Product"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Like recorded successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Like" },
                "example": {
                  "_id": "64b1f4c2e4b0f5a3d8e9g333",
                  "user": "64b1f4c2e4b0f5a3d8e9d000",
                  "likeable": "64b1f4c2e4b0f5a3d8e9a123",
                  "types": "Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing id/type or invalid type value.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "likeSchema validation failed: types: `InvalidType` is not a valid enum value for path `types`." }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Likes"],
        "summary": "Get likes for a product or category",
        "description": "Returns all likes for a specific product or category identified by its ObjectId and type.",
        "security": [{ "JWT": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "example": "64b1f4c2e4b0f5a3d8e9a123" },
            "description": "The MongoDB ObjectId of the product or category."
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Product", "Category"],
              "example": "Product"
            },
            "description": "The type of the entity to retrieve likes for."
          }
        ],
        "responses": {
          "200": {
            "description": "Likes for the specified entity returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Like" }
                },
                "example": [
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9g333",
                    "user": "64b1f4c2e4b0f5a3d8e9d000",
                    "likeable": "64b1f4c2e4b0f5a3d8e9a123",
                    "types": "Product"
                  },
                  {
                    "_id": "64b1f4c2e4b0f5a3d8e9g334",
                    "user": "64b1f4c2e4b0f5a3d8e9d001",
                    "likeable": "64b1f4c2e4b0f5a3d8e9a123",
                    "types": "Product"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Bad request - missing or invalid id/type query parameters.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "id and type query parameters are required." }
              }
            }
          },
          "401": {
            "description": "Unauthorized - JWT token is missing or invalid.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Token is missing/unauthorized" }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "message": "Internal Server Error, Please try again later." }
              }
            }
          }
        }
      }
    },
...
  }
}